# db: p.stmt and query will now reconnect after disconnects