<!-- Remove space and place 'x' mark between square [] brackets or click the checkbox after saving to affirm the following points: -->
<!-- (it should look like this: - [x] I have ...) -->
**_I affirm:_**
- [x] I understand that if I do not agree to the following points by completing the checkboxes my PR will be ignored.
- [x] I understand I should leave resolving conversations to the LandSandBoat team so that reviewers won't miss what was said.
- [x] I have read and understood the [Contributing Guide](https://github.com/LandSandBoat/server/blob/base/CONTRIBUTING.md) and the [Code of Conduct](https://github.com/LandSandBoat/server/blob/base/CODE_OF_CONDUCT.md).
- [x] I have _**tested my code and the things my code has changed**_ since the last commit in the PR and will test after any later commits.

## What does this pull request do?

Not my finest design, but this is the general layout:
```cpp

        try
        {
            return operation();
        }
        catch (const std::exception& e)
        {
            const auto str = fmt::format("{}", e.what());
            if (str.find("Lost connection") == std::string::npos) // If not a connection issue, report the error
            {
                ShowError("Query Failed: %s", rawQuery.c_str());
                ShowError(e.what());
                return nullptr;
            }
        }

        ShowInfo("Connection lost, re-establishing connection and retrying query");
        state.reset();

        try
        {
            return operation();
        }
        catch (const std::exception& e)
        {
            ShowError("Query Failed: %s", rawQuery.c_str());
            ShowError(e.what());
        }

        return nullptr;
```