@zach2good 2024-12-21
Updated the logic here:
- We no longer ping the db before _every_ query, this should result in a speedup. I'm quite hyped about this, so I'll go collect some numbers.
- We try every query. If it fails due to "lost connection" or "server went away" we'll rebuild the connection and try again n times (there's a setting for that)
- If _at any stage_ we try to build a connection to the db and it fails - the server needs to die. This is an ABSOLUTE MEGA FAILURE STATE and we absolutely don't want to try and recover from it. Sorry not sorry.