@zach2good 2024-12-21
Old system (with pinging, doesn't correctly safety check or reconnect):
```
db_perf_test
> db_perf_test queries took 173ms
> db_perf_test prepared statements took 181ms
db_perf_test
> db_perf_test queries took 158ms
> db_perf_test prepared statements took 158ms
db_perf_test
> db_perf_test queries took 243ms
> db_perf_test prepared statements took 176ms
db_perf_test
> db_perf_test queries took 160ms
> db_perf_test prepared statements took 166ms
db_perf_test
> db_perf_test queries took 160ms
> db_perf_test prepared statements took 154ms
```

New system (no pinging, will repair connection if it goes away):
```
db_perf_test
> db_perf_test queries took 135ms
> db_perf_test prepared statements took 136ms
db_perf_test
> db_perf_test queries took 130ms
> db_perf_test prepared statements took 135ms
db_perf_test
> db_perf_test queries took 132ms
> db_perf_test prepared statements took 149ms
db_perf_test
> db_perf_test queries took 151ms
> db_perf_test prepared statements took 129ms
db_perf_test
> db_perf_test queries took 139ms
> db_perf_test prepared statements took 133ms
```

Comparison, as generated by ChatGPT:

# Performance Comparison: Old vs. New DB System

Below is a brief, high-level comparison of the old system vs. the new system and how queries compare to prepared statements based on the data you provided.

---

## 1. Data Recap

### Old System

*(Has pinging, doesn’t always safety-check or reconnect properly)*

| Run | Queries (ms) | Prepared (ms) |
|-----|--------------|---------------|
| 1   | 173          | 181           |
| 2   | 158          | 158           |
| 3   | 243          | 176           |
| 4   | 160          | 166           |
| 5   | 160          | 154           |

- **Average (Queries)**  
  \[
    \frac{173 + 158 + 243 + 160 + 160}{5} = 178.8\,\text{ms}
  \]

- **Average (Prepared)**  
  \[
    \frac{181 + 158 + 176 + 166 + 154}{5} = 167\,\text{ms}
  \]

### New System

*(No pinging, automatically repairs connection if it fails)*

| Run | Queries (ms) | Prepared (ms) |
|-----|--------------|---------------|
| 1   | 135          | 136           |
| 2   | 130          | 135           |
| 3   | 132          | 149           |
| 4   | 151          | 129           |
| 5   | 139          | 133           |

- **Average (Queries)**  
  \[
    \frac{135 + 130 + 132 + 151 + 139}{5} = 137.4\,\text{ms}
  \]

- **Average (Prepared)**  
  \[
    \frac{136 + 135 + 149 + 129 + 133}{5} = 136.4\,\text{ms}
  \]

---

## 2. Observations

1. **New System is Faster Overall**  
   - Old system queries: **178.8 ms** on average  
   - New system queries: **137.4 ms** on average  
   - Old system prepared statements: **167 ms** on average  
   - New system prepared statements: **136.4 ms** on average  
   <br>
   The new system is roughly 20–40 ms faster on average for both queries and prepared statements.

2. **Prepared vs. Queries**  
   - **Old system**: Prepared statements (167 ms) are faster than raw queries (178.8 ms).  
   - **New system**: Prepared statements (136.4 ms) vs. queries (137.4 ms) are very close. Sometimes queries are slightly faster; sometimes prepared is.

3. **Consistency and Variance**  
   - The old system occasionally had large spikes (e.g., **243 ms** for queries).  
   - The new system’s worst outliers are less drastic (e.g., **151 ms** for queries, **149 ms** for prepared statements).  
   - Likely, the old system’s overhead from “pinging” and poor reconnection handling introduced more variance. The new system’s auto-repair logic provides more stable performance.

---

## 3. Possible Explanations

1. **Reduced Overhead**  
   Eliminating extra ping logic in the new system likely reduces overhead, trimming a few milliseconds off each test.

2. **Improved Connection Handling**  
   The old system didn’t properly handle reconnections, resulting in occasional spikes. The new system’s “repair if dropped” approach is more efficient and robust.

3. **Small Differences for Prepared Statements**  
   - Prepared statements help if you reuse the same query repeatedly.  
   - In this short test scenario, the differences can be small or overshadowed by network/driver overhead.  
   - The main gains in the new system come from improved connection handling rather than the query execution strategy itself.

---

## 4. Key Takeaways

- **Significant Speedup**: The new system consistently outperforms the old system for both raw queries and prepared statements.  
- **Lower Variability**: Performance in the new system is more tightly clustered, which points to better connection stability and less overhead.  
- **Prepared vs. Queries**: Prepared statements are marginally faster or on par with queries in these tests. The biggest win is still from the new connection handling and eliminating ping overhead.