@zach2good 2024-12-21
So, roughly:

Queries: ~23% faster
Prepared statements: ~18% faster

On our slowest, most frequent, and most blocking system.