@zach2good 2024-12-21
and for fairness, I built the same test harness for the old system, which has: zero thread safety, zero convenience, entirely global state, no resilience or self repair, no prepared statements, server-based string escaping, etc.

```
db_perf_test
> db_perf_test queries took 133ms
db_perf_test
> db_perf_test queries took 138ms
db_perf_test
> db_perf_test queries took 123ms
db_perf_test
> db_perf_test queries took 130ms
db_perf_test
> db_perf_test queries took 147ms
```

Pretty much the same as the new system, with slightly more variance?