@zach2good 2024-12-21
This pretty much wraps up and confirms that the new db system is ready for primetime - it does everything the old system does, as fast, with many many more bells and whistles.