@StabbyCutyou 2024-12-21
Apologies as I'm coming in post-merge but, on the approach to retry queries N times, is there a risk that a write is accepted, then an error occurs as the response is being sent to the client (a read timeout, a socket error of some kind, etc), then you retry the write and end up in a bad state? I can see you're scoping it to a limited set of exceptions, but wondering if it's possible for this case to appear with one of those errors.

Fairly odd edge case to consider, but just wondering if you think that might be an issue to look out for.

I'm also curious if you've considered a connection pooling approach, as typically they'll handle always ensuring the caller gets a clean connection (and increase throughput at the cost of a bit of general latency) but that is probably a different discussion that you've already had somewhere else.