@zach2good 2024-12-21
In terms of what might affect the game, players, and game state data: I don't think we need to worry about multiple query retry. If we try to insert twice we'll be inserting to the same item slot, if we try to delete twice the second delete will be moot, etc.

Connection pooling is a tough one, it's on my list of things to do. The MariaDB connectors we use (C and C++ versions) don't provide all the flags and options they claim to - including pooling, auto-reconnect, etc. This is why I've done the reconnection logic manually. At some point I'll handle doing connection pooling manually, but that'll likely come after the other db, ASIO, and worker thread work I have upcoming.